refactor: 재사용성을 높이기 위해 중복되는 api 요청 메소드 통합
