--- src/app/_components/common/ProfileImage.tsx
@@ -1,12 +1,0 @@
-'use client'
-
-const ProfileImage = () => {
-    //Fetch 매칭된 상대 데이터 가져오기
-    return (
-        <div>
-
-        </div>
-    )
-}
-
-export default ProfileImage;
