+++ src/app/_components/signupsteps/Step2.tsx
@@ -1,46 +1,53 @@
-import { useState, useEffect } from "react";
-import StepIndicator from '../../_components/stepIndicator';
+import { useState, useEffect } from 'react';
+import StepIndicator from '../stepIndicator';
 import NavigationButtons from './navigationbuttons/NavigationButtons';
 
 interface Step2Props {
     nextStep: () => void;
     prevStep: () => void;
-    updateFormData: (data: { birthday: string }) => void;
+    updateFormData: (data: { gender: string }) => void;
     formData: any;
 }
 
 const Step2: React.FC<Step2Props> = ({ nextStep, prevStep, updateFormData, formData }) => {
-    const [birthday, setBirthday] = useState(formData.birthday || "");
+    const [gender, setGender] = useState<string | null>(formData.gender || null);
 
     useEffect(() => {
-        setBirthday(formData.birthday || "");
-    }, [formData.birthday]);
+        setGender(formData.gender || null);
+    }, [formData.gender]);
 
     const handleSubmit = (event: React.FormEvent) => {
         event.preventDefault();
-        if (!birthday) {
-            alert("생일을 입력해주세요.");
-            return;
+        if (gender) {
+            updateFormData({ gender });
+            nextStep();
+        } else {
+            alert("성별을 선택해주세요.");
         }
-        updateFormData({ birthday });
-        nextStep();
     };
 
     return (
         <div className="flex items-center justify-center min-h-screen">
             <div className="w-full max-w-xl p-12 bg-white shadow-md rounded-lg border-4 border-red-300">
-                <StepIndicator currentStep={1} />
+                <StepIndicator currentStep={2} />
                 <form onSubmit={handleSubmit} className="space-y-6">
-                    <label htmlFor="birthday" className="block text-4xl text-center mb-40">생일은 언제인가요?</label>
-                    <input
-                        type="date"
-                        id="birthday"
-                        name="birthday"
-                        required
-                        value={birthday}
-                        onChange={(e) => setBirthday(e.target.value)}
-                        className="text-center text-xl w-full border-b-2 border-black focus:outline-none focus:border-red-400"
-                    />
+                    <label className="block text-4xl text-center mb-40">성별은 무엇인가요?</label>
+                    <div className="flex justify-center space-x-4">
+                        <button
+                            type="button"
+                            onClick={() => setGender('남성')}
+                            className={`py-2 px-6 border-4 rounded-lg text-2xl ${gender === '남성' ? 'border-red-500 bg-red-300 text-white' : 'border-red-300 bg-white text-black'}`}
+                        >
+                            남성
+                        </button>
+                        <button
+                            type="button"
+                            onClick={() => setGender('여성')}
+                            className={`py-2 px-6 border-4 rounded-lg text-2xl ${gender === '여성' ? 'border-red-500 bg-red-300 text-white' : 'border-red-300 bg-white text-black'}`}
+                        >
+                            여성
+                        </button>
+                    </div>
                     <NavigationButtons onNext={handleSubmit} onPrevious={prevStep} />
                 </form>
             </div>
